Move test to the base trait
